+++ CHANGELOG.md
@@ -7,6 +7,10 @@ References and Resolving:
 1.  Support `URIQualifiedName` references, both the namespace part (`Q{...}`) and the function or variable `NCName`.
 1.  Support navigating to Java classes specified by `java:` and unprefixed classpath URIs (BaseX, eXist-db, Saxon).
 
+IntelliJ Integration:
+
+1.  Display the namespace signature and module path when holding Ctrl over a module namespace URI or prefix.
+
 ## 1.5.2 - 2019-11-19
 
 Code Completion: